#70 Move traits
