+++ app/Concerns/Companies/DetectsUKPhoneNumbers.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Concerns;
+namespace App\Concerns\Companies;
 
 /**
  * Detect UK phone numbers.
