+++ app/Concerns/Companies/FormatsGenericPhoneNumbers.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Concerns;
+namespace App\Concerns\Companies;
 
 /**
  * Generic phone number formatting helpers.
