+++ app/Concerns/Companies/FormatsInternationalPhoneNumbers.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Concerns;
+namespace App\Concerns\Companies;
 
 /**
  * International phone number formatting helpers.
