+++ app/Concerns/Companies/FormatsUKPhoneNumbers.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Concerns;
+namespace App\Concerns\Companies;
 
 /**
  * UK phone number formatting helpers.
