+++ app/Concerns/Companies/FormatsUSPhoneNumbers.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Concerns;
+namespace App\Concerns\Companies;
 
 /**
  * US phone number formatting helpers.
