+++ app/Concerns/Companies/HasCompanyContactHelpers.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Concerns;
+namespace App\Concerns\Companies;
 
 /**
  * @property string $first_name
