+++ app/Concerns/Companies/HasCompanyHelpers.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Concerns;
+namespace App\Concerns\Companies;
 
 /**
  * @property string|null $address
