+++ app/Concerns/Companies/HasCompanyPhoneStateHelpers.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Concerns;
+namespace App\Concerns\Companies;
 
 /**
  * Company phone state helper methods.
