+++ app/Concerns/Companies/HasCompanyPhoneTypeHelpers.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Concerns;
+namespace App\Concerns\Companies;
 
 use App\Models\CompanyPhone;
