+++ app/Concerns/Companies/SplitsUKPhoneNumbers.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Concerns;
+namespace App\Concerns\Companies;
 
 /**
  * Split UK phone numbers.
