+++ app/Concerns/Pipelines/HasPipelineScopes.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Concerns;
+namespace App\Concerns\Pipelines;
 
 use App\Models\Pipeline;
 use Illuminate\Database\Eloquent\Builder;
