+++ app/Concerns/Users/HasUserScopes.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Concerns;
+namespace App\Concerns\Users;
 
 use Illuminate\Database\Eloquent\Builder;
